@@ -40,15 +40,15 @@
         "registry": "https://repository.hybris.com/api/npm/npm-repository/"
     },
     "dependencies": {
-        "@angular/animations": "21.0.6",
-        "@angular/cdk": "21.0.5",
-        "@angular/common": "21.0.6",
-        "@angular/compiler": "21.0.6",
-        "@angular/core": "21.0.6",
-        "@angular/forms": "21.0.6",
-        "@angular/platform-browser": "21.0.6",
-        "@angular/platform-browser-dynamic": "21.0.6",
-        "@angular/router": "21.0.6",
+        "@angular/animations": "21.1.0",
+        "@angular/cdk": "21.1.0",
+        "@angular/common": "21.1.0",
+        "@angular/compiler": "21.1.0",
+        "@angular/core": "21.1.0",
+        "@angular/forms": "21.1.0",
+        "@angular/platform-browser": "21.1.0",
+        "@angular/platform-browser-dynamic": "21.1.0",
+        "@angular/router": "21.1.0",
         "@fundamental-styles/cx": "0.40.1",
         "@nx/angular": "22.3.3",
         "@sap-theming/theming-base-content": "11.32.3",
@@ -84,12 +84,12 @@
     },
     "devDependencies": {
         "@actions/core": "1.10.1",
-        "@angular-devkit/build-angular": "21.0.4",
-        "@angular-devkit/core": "21.0.4",
-        "@angular-devkit/schematics": "21.0.4",
-        "@angular/cli": "21.0.4",
-        "@angular/compiler-cli": "21.0.6",
-        "@angular/language-service": "21.0.6",
+        "@angular-devkit/build-angular": "21.1.0",
+        "@angular-devkit/core": "21.1.0",
+        "@angular-devkit/schematics": "21.1.0",
+        "@angular/cli": "21.1.0",
+        "@angular/compiler-cli": "21.1.0",
+        "@angular/language-service": "21.1.0",
         "@commitlint/cli": "18.6.1",
         "@commitlint/config-conventional": "18.6.1",
         "@nx/devkit": "22.3.3",
@@ -100,7 +100,7 @@
         "@nx/plugin": "22.3.3",
         "@nx/workspace": "22.3.3",
         "@sap-ui/common-css": "0.40.1",
-        "@schematics/angular": "21.0.4",
+        "@schematics/angular": "21.1.0",
         "@swc-node/register": "1.9.2",
         "@swc/cli": "0.6.0",
         "@swc/core": "1.5.7",
